moved publisher related classes to publisher package added unit tests for the publisher Factory
