+++ packager/src/main/java/io/cdap/hub/Tool.java
@@ -19,6 +19,9 @@ package io.cdap.hub;
 import com.google.common.base.Splitter;
 import com.google.common.io.BaseEncoding;
 import com.google.common.primitives.Longs;
+import io.cdap.hub.publisher.Publisher;
+import io.cdap.hub.publisher.PublisherFactory;
+import io.cdap.hub.publisher.S3Publisher;
 import org.apache.commons.cli.BasicParser;
 import org.apache.commons.cli.CommandLine;
 import org.apache.commons.cli.CommandLineParser;
