+++ packager/src/main/java/io/cdap/hub/publisher/GCSPublisher.java
@@ -14,24 +14,23 @@
  * the License.
  */
 
-package io.cdap.hub;
+package io.cdap.hub.publisher;
 
-import com.amazonaws.ClientConfiguration;
-import com.amazonaws.Protocol;
-import com.amazonaws.auth.BasicAWSCredentials;
-import com.amazonaws.services.cloudfront.AmazonCloudFrontClient;
-import com.amazonaws.services.cloudfront.model.CreateInvalidationRequest;
-import com.amazonaws.services.cloudfront.model.InvalidationBatch;
-import com.amazonaws.services.cloudfront.model.Paths;
-import com.amazonaws.services.s3.AmazonS3Client;
-import com.amazonaws.services.s3.model.CannedAccessControlList;
-import com.amazonaws.services.s3.model.ObjectMetadata;
-import com.amazonaws.services.s3.model.PutObjectRequest;
-import com.amazonaws.services.s3.model.S3ObjectSummary;
+import com.google.api.gax.paging.Page;
+import com.google.cloud.storage.Blob;
+import com.google.cloud.storage.BlobId;
+import com.google.cloud.storage.BlobInfo;
+import com.google.cloud.storage.Bucket;
+import com.google.cloud.storage.Storage;
+import com.google.cloud.storage.StorageOptions;
+import com.google.common.collect.ImmutableMap;
 import com.google.common.hash.Hashing;
 import com.google.common.io.BaseEncoding;
 import com.google.common.io.Files;
 import com.google.common.net.MediaType;
+import io.cdap.hub.Hub;
+import io.cdap.hub.Package;
+import io.cdap.hub.SignedFile;
 import io.cdap.hub.spec.CategoryMeta;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -41,45 +40,40 @@ import java.io.IOException;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
+
 import javax.activation.FileTypeMap;
 import javax.activation.MimetypesFileTypeMap;
 import javax.annotation.Nullable;
 
 /**
- * Publish packages to S3.
+ * Publish packages to GCS bucket.
  */
-public class S3Publisher implements Publisher {
-  private static final Logger LOG = LoggerFactory.getLogger(S3Publisher.class);
+public class GCSPublisher implements Publisher {
+  private static final Logger LOG = LoggerFactory.getLogger(GCSPublisher.class);
   private static final FileTypeMap fileTypeMap = MimetypesFileTypeMap.getDefaultFileTypeMap();
-  private final AmazonS3Client s3Client;
-  @Nullable
-  private final AmazonCloudFrontClient cfClient;
+  private Storage storage;
+  private final String projectId;
   private final String bucket;
   private final String prefix;
-  @Nullable
-  private final String cfDistribution;
+
   private final boolean forcePush;
   private final boolean dryrun;
-  private final Set<String> whitelist;
   private final Set<String> updatedKeys;
 
-  private S3Publisher(AmazonS3Client s3Client, @Nullable AmazonCloudFrontClient cfClient,
-                      String bucket, String prefix, @Nullable String cfDistribution,
-                      boolean forcePush, boolean dryrun, Set<String> whitelist) {
-    this.s3Client = s3Client;
-    this.cfClient = cfClient;
+  private GCSPublisher(String projectId, String bucket, String prefix, boolean dryrun, boolean forcePush) {
+    this.projectId = projectId;
     this.bucket = bucket;
     this.prefix = prefix;
-    this.cfDistribution = cfDistribution;
-    this.forcePush = forcePush;
     this.dryrun = dryrun;
-    this.whitelist = whitelist;
+    this.forcePush = forcePush;
     this.updatedKeys = new HashSet<>();
   }
 
   @Override
   public void publish(Hub hub) throws Exception {
+    this.storage = StorageOptions.newBuilder().setProjectId(projectId).build().getService();
     updatedKeys.clear();
+
     List<Package> packages = hub.getPackages();
     for (Package pkg : packages) {
       publishPackage(pkg);
@@ -91,21 +85,6 @@ public class S3Publisher implements Publisher {
     putFilesIfChanged(prefix + "/", hub.getPackageCatalog());
     LOG.info("Publishing category catalog");
     putFilesIfChanged(prefix + "/", hub.getCategoryCatalog());
-
-    if (cfClient != null && !updatedKeys.isEmpty()) {
-      CreateInvalidationRequest invalidationRequest = new CreateInvalidationRequest()
-        .withDistributionId(cfDistribution)
-        .withInvalidationBatch(
-          new InvalidationBatch()
-            .withPaths(new Paths().withItems(updatedKeys).withQuantity(updatedKeys.size()))
-            .withCallerReference(String.valueOf(System.currentTimeMillis())));
-      if (!dryrun) {
-        LOG.info("Invalidating cloudfront objects {}", updatedKeys);
-        cfClient.createInvalidation(invalidationRequest);
-      } else {
-        LOG.info("dryrun - would have invalidated cloudfront objects {}", updatedKeys);
-      }
-    }
   }
 
   private void publishPackage(Package pkg) throws Exception {
@@ -121,13 +100,21 @@ public class S3Publisher implements Publisher {
     for (SignedFile file : pkg.getFiles()) {
       putFilesIfChanged(keyPrefix, file.getFile(), file.getSignature());
     }
-    for (S3ObjectSummary objectSummary : s3Client.listObjects(bucket, keyPrefix).getObjectSummaries()) {
-      String objectKey = objectSummary.getKey();
+
+    Bucket bucket = storage.get(this.bucket);
+    Page<Blob> blobs =
+            bucket.list(
+                    Storage.BlobListOption.prefix(keyPrefix),
+                    Storage.BlobListOption.currentDirectory());
+
+    for (Blob blob : blobs.iterateAll()) {
+
+      String objectKey = blob.getName();
       String name = objectKey.substring(keyPrefix.length());
       if (!pkg.getFileNames().contains(name)) {
         if (!dryrun) {
           LOG.info("Deleting object {} from s3 since it does not exist in the package anymore.", objectKey);
-          s3Client.deleteObject(bucket, objectKey);
+          storage.delete(blob.getBlobId());
         } else {
           LOG.info("dryrun - would have deleted {} from s3 since it does not exist in the package anymore.", objectKey);
         }
@@ -152,19 +139,25 @@ public class S3Publisher implements Publisher {
     }
   }
 
-  // check if the file on s3 has a different md5 or the file length.
+  // check if the file in the gcs bucket has a different md5 or the file length.
   private boolean shouldPush(String keyPrefix, File file) throws IOException {
     if (forcePush) {
       return true;
     }
     String key = keyPrefix + file.getName();
-    if (s3Client.doesObjectExist(bucket, key)) {
-      ObjectMetadata existingMeta = s3Client.getObjectMetadata(bucket, key);
+
+    Bucket bucket = storage.get(this.bucket);
+    Page<Blob> blobs =
+            bucket.list(
+                    Storage.BlobListOption.prefix(key),
+                    Storage.BlobListOption.currentDirectory());
+    if (blobs.hasNextPage()) {
+      Blob blob = blobs.getNextPage().getValues().iterator().next();
+      long existingContentFileLength = blob.getSize();
       long fileLength = file.length();
       String md5Hex = BaseEncoding.base16().encode(Files.hash(file, Hashing.md5()).asBytes());
-      if (existingMeta != null &&
-        existingMeta.getContentLength() == fileLength &&
-        existingMeta.getETag() != null && existingMeta.getETag().equalsIgnoreCase(md5Hex)) {
+      if (existingContentFileLength == fileLength &&
+        blob.getEtag() != null && blob.getEtag().equalsIgnoreCase(md5Hex)) {
         LOG.info("{} has not changed, skipping upload to S3.", file);
         return false;
       }
@@ -191,65 +184,43 @@ public class S3Publisher implements Publisher {
       default:
         contentType = fileTypeMap.getContentType(file);
     }
-    ObjectMetadata newMeta = new ObjectMetadata();
-    newMeta.setContentType(contentType);
+
     String key = keyPrefix + file.getName();
-    PutObjectRequest request = new PutObjectRequest(bucket, key, file)
-      .withCannedAcl(CannedAccessControlList.PublicRead)
-      .withMetadata(newMeta);
+
     if (!dryrun) {
       LOG.info("put file {} into s3 with key {}", file, key);
-      s3Client.putObject(request);
+      BlobId blobId = BlobId.of(bucket, key);
+      BlobInfo blobInfo = BlobInfo.newBuilder(blobId)
+              .setMetadata(ImmutableMap.of("Content-Type", contentType))
+              .build();
+      storage.create(blobInfo, Files.toByteArray(file));
     } else {
       LOG.info("dryrun - would have put file {} into s3 with key {}", file, key);
     }
     updatedKeys.add("/" + key);
   }
 
-  public static Builder builder(String s3Bucket, String s3AccessKey, String s3SecretKey) {
-    return new Builder(s3Bucket, s3AccessKey, s3SecretKey);
+  public static Builder builder(String projectId, String bucket) {
+    return new Builder(projectId, bucket);
   }
 
   /**
    * Builder to create the S3Publisher.
    */
   public static class Builder {
-    private final String s3Bucket;
-    private final String s3AccessKey;
-    private final String s3SecretKey;
+
+    private final String projectId;
+    private final String bucket;
     private String prefix;
-    private String cfDistribution;
-    private String cfAccessKey;
-    private String cfSecretKey;
     private boolean forcePush;
     private boolean dryrun;
-    private int timeout;
-    private Set<String> whitelist;
 
-    public Builder(String s3Bucket, String s3AccessKey, String s3SecretKey) {
-      this.s3Bucket = s3Bucket;
-      this.s3AccessKey = s3AccessKey;
-      this.s3SecretKey = s3SecretKey;
-      forcePush = false;
-      dryrun = false;
-      timeout = 30;
-      prefix = "";
-      whitelist = new HashSet<>();
-    }
-
-    public Builder setCloudfrontDistribution(String distribution) {
-      this.cfDistribution = distribution;
-      return this;
-    }
-
-    public Builder setCloudfrontAccessKey(String accessKey) {
-      this.cfAccessKey = accessKey;
-      return this;
-    }
-
-    public Builder setCloudfrontSecretKey(String secretKey) {
-      this.cfSecretKey = secretKey;
-      return this;
+    public Builder(String projectId, String bucket) {
+      this.projectId = projectId;
+      this.bucket = bucket;
+      this.forcePush = false;
+      this.dryrun = false;
+      this.prefix = "";
     }
 
     public Builder setPrefix(String prefix) {
@@ -267,33 +238,8 @@ public class S3Publisher implements Publisher {
       return this;
     }
 
-    public Builder setTimeout(int timeout) {
-      this.timeout = timeout;
-      return this;
-    }
-
-    public Builder setWhitelist(Set<String> whitelist) {
-      this.whitelist = whitelist;
-      return this;
-    }
-
-    public S3Publisher build() {
-      ClientConfiguration clientConf = new ClientConfiguration()
-        .withProtocol(Protocol.HTTPS)
-        .withSocketTimeout(timeout * 1000);
-
-      AmazonS3Client s3Client = new AmazonS3Client(new BasicAWSCredentials(s3AccessKey, s3SecretKey), clientConf);
-
-      AmazonCloudFrontClient cfClient = null;
-      if (cfDistribution != null) {
-        if (cfAccessKey == null || cfSecretKey == null) {
-          throw new IllegalArgumentException(
-            "When specifying a cloudfront distribution, must also specify a cloudfront access key and secret key.");
-        }
-        cfClient = new AmazonCloudFrontClient(new BasicAWSCredentials(cfAccessKey, cfSecretKey), clientConf);
-      }
-
-      return new S3Publisher(s3Client, cfClient, s3Bucket, prefix, cfDistribution, forcePush, dryrun, whitelist);
+    public GCSPublisher build() {
+      return new GCSPublisher(projectId, bucket, prefix, dryrun, forcePush);
     }
   }
 }
