+++ packager/src/main/java/io/cdap/hub/publisher/Publisher.java
@@ -1,0 +1,33 @@
+/*
+ * Copyright © 2016 Cask Data, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not
+ * use this file except in compliance with the License. You may obtain a copy of
+ * the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+ * License for the specific language governing permissions and limitations under
+ * the License.
+ */
+
+package io.cdap.hub.publisher;
+
+import io.cdap.hub.Hub;
+
+/**
+ * Publishes the Hub.
+ */
+public interface Publisher {
+
+  /**
+   * Publish the hub.
+   *
+   * @param hub the hub to publish
+   * @throws Exception if there was an error publishing
+   */
+  void publish(Hub hub) throws Exception;
+}
