+++ packager/src/main/java/io/cdap/hub/publisher/S3Publisher.java
@@ -14,20 +14,27 @@
  * the License.
  */
 
-package io.cdap.hub;
+package io.cdap.hub.publisher;
 
-import com.google.api.gax.paging.Page;
-import com.google.cloud.storage.Blob;
-import com.google.cloud.storage.BlobId;
-import com.google.cloud.storage.BlobInfo;
-import com.google.cloud.storage.Bucket;
-import com.google.cloud.storage.Storage;
-import com.google.cloud.storage.StorageOptions;
-import com.google.common.collect.ImmutableMap;
+import com.amazonaws.ClientConfiguration;
+import com.amazonaws.Protocol;
+import com.amazonaws.auth.BasicAWSCredentials;
+import com.amazonaws.services.cloudfront.AmazonCloudFrontClient;
+import com.amazonaws.services.cloudfront.model.CreateInvalidationRequest;
+import com.amazonaws.services.cloudfront.model.InvalidationBatch;
+import com.amazonaws.services.cloudfront.model.Paths;
+import com.amazonaws.services.s3.AmazonS3Client;
+import com.amazonaws.services.s3.model.CannedAccessControlList;
+import com.amazonaws.services.s3.model.ObjectMetadata;
+import com.amazonaws.services.s3.model.PutObjectRequest;
+import com.amazonaws.services.s3.model.S3ObjectSummary;
 import com.google.common.hash.Hashing;
 import com.google.common.io.BaseEncoding;
 import com.google.common.io.Files;
 import com.google.common.net.MediaType;
+import io.cdap.hub.Hub;
+import io.cdap.hub.Package;
+import io.cdap.hub.SignedFile;
 import io.cdap.hub.spec.CategoryMeta;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -37,31 +44,39 @@ import java.io.IOException;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
-
 import javax.activation.FileTypeMap;
 import javax.activation.MimetypesFileTypeMap;
 import javax.annotation.Nullable;
 
 /**
- * Publish packages to GCS bucket.
+ * Publish packages to S3.
  */
-public class GCSPublisher implements Publisher {
-  private static final Logger LOG = LoggerFactory.getLogger(GCSPublisher.class);
+public class S3Publisher implements Publisher {
+  private static final Logger LOG = LoggerFactory.getLogger(S3Publisher.class);
   private static final FileTypeMap fileTypeMap = MimetypesFileTypeMap.getDefaultFileTypeMap();
-  private final Storage storage;
+  private final AmazonS3Client s3Client;
+  @Nullable
+  private final AmazonCloudFrontClient cfClient;
   private final String bucket;
   private final String prefix;
-
+  @Nullable
+  private final String cfDistribution;
   private final boolean forcePush;
   private final boolean dryrun;
+  private final Set<String> whitelist;
   private final Set<String> updatedKeys;
 
-  private GCSPublisher(String projectId, String bucket, String prefix, boolean dryrun, boolean forcePush) {
-    this.storage = StorageOptions.newBuilder().setProjectId(projectId).build().getService();
+  private S3Publisher(AmazonS3Client s3Client, @Nullable AmazonCloudFrontClient cfClient,
+                      String bucket, String prefix, @Nullable String cfDistribution,
+                      boolean forcePush, boolean dryrun, Set<String> whitelist) {
+    this.s3Client = s3Client;
+    this.cfClient = cfClient;
     this.bucket = bucket;
     this.prefix = prefix;
-    this.dryrun = dryrun;
+    this.cfDistribution = cfDistribution;
     this.forcePush = forcePush;
+    this.dryrun = dryrun;
+    this.whitelist = whitelist;
     this.updatedKeys = new HashSet<>();
   }
 
@@ -79,6 +94,21 @@ public class GCSPublisher implements Publisher {
     putFilesIfChanged(prefix + "/", hub.getPackageCatalog());
     LOG.info("Publishing category catalog");
     putFilesIfChanged(prefix + "/", hub.getCategoryCatalog());
+
+    if (cfClient != null && !updatedKeys.isEmpty()) {
+      CreateInvalidationRequest invalidationRequest = new CreateInvalidationRequest()
+        .withDistributionId(cfDistribution)
+        .withInvalidationBatch(
+          new InvalidationBatch()
+            .withPaths(new Paths().withItems(updatedKeys).withQuantity(updatedKeys.size()))
+            .withCallerReference(String.valueOf(System.currentTimeMillis())));
+      if (!dryrun) {
+        LOG.info("Invalidating cloudfront objects {}", updatedKeys);
+        cfClient.createInvalidation(invalidationRequest);
+      } else {
+        LOG.info("dryrun - would have invalidated cloudfront objects {}", updatedKeys);
+      }
+    }
   }
 
   private void publishPackage(Package pkg) throws Exception {
@@ -94,21 +124,13 @@ public class GCSPublisher implements Publisher {
     for (SignedFile file : pkg.getFiles()) {
       putFilesIfChanged(keyPrefix, file.getFile(), file.getSignature());
     }
-
-    Bucket bucket = storage.get(this.bucket);
-    Page<Blob> blobs =
-            bucket.list(
-                    Storage.BlobListOption.prefix(keyPrefix),
-                    Storage.BlobListOption.currentDirectory());
-
-    for (Blob blob : blobs.iterateAll()) {
-
-      String objectKey = blob.getName();
+    for (S3ObjectSummary objectSummary : s3Client.listObjects(bucket, keyPrefix).getObjectSummaries()) {
+      String objectKey = objectSummary.getKey();
       String name = objectKey.substring(keyPrefix.length());
       if (!pkg.getFileNames().contains(name)) {
         if (!dryrun) {
           LOG.info("Deleting object {} from s3 since it does not exist in the package anymore.", objectKey);
-          storage.delete(blob.getBlobId());
+          s3Client.deleteObject(bucket, objectKey);
         } else {
           LOG.info("dryrun - would have deleted {} from s3 since it does not exist in the package anymore.", objectKey);
         }
@@ -133,25 +155,19 @@ public class GCSPublisher implements Publisher {
     }
   }
 
-  // check if the file in the gcs bucket has a different md5 or the file length.
+  // check if the file on s3 has a different md5 or the file length.
   private boolean shouldPush(String keyPrefix, File file) throws IOException {
     if (forcePush) {
       return true;
     }
     String key = keyPrefix + file.getName();
-
-    Bucket bucket = storage.get(this.bucket);
-    Page<Blob> blobs =
-            bucket.list(
-                    Storage.BlobListOption.prefix(key),
-                    Storage.BlobListOption.currentDirectory());
-    if (blobs.hasNextPage()) {
-      Blob blob = blobs.getNextPage().getValues().iterator().next();
-      long existingContentFileLength = blob.getSize();
+    if (s3Client.doesObjectExist(bucket, key)) {
+      ObjectMetadata existingMeta = s3Client.getObjectMetadata(bucket, key);
       long fileLength = file.length();
       String md5Hex = BaseEncoding.base16().encode(Files.hash(file, Hashing.md5()).asBytes());
-      if (existingContentFileLength == fileLength &&
-        blob.getEtag() != null && blob.getEtag().equalsIgnoreCase(md5Hex)) {
+      if (existingMeta != null &&
+        existingMeta.getContentLength() == fileLength &&
+        existingMeta.getETag() != null && existingMeta.getETag().equalsIgnoreCase(md5Hex)) {
         LOG.info("{} has not changed, skipping upload to S3.", file);
         return false;
       }
@@ -178,43 +194,65 @@ public class GCSPublisher implements Publisher {
       default:
         contentType = fileTypeMap.getContentType(file);
     }
-
+    ObjectMetadata newMeta = new ObjectMetadata();
+    newMeta.setContentType(contentType);
     String key = keyPrefix + file.getName();
-
+    PutObjectRequest request = new PutObjectRequest(bucket, key, file)
+      .withCannedAcl(CannedAccessControlList.PublicRead)
+      .withMetadata(newMeta);
     if (!dryrun) {
       LOG.info("put file {} into s3 with key {}", file, key);
-      BlobId blobId = BlobId.of(bucket, key);
-      BlobInfo blobInfo = BlobInfo.newBuilder(blobId)
-              .setMetadata(ImmutableMap.of("Content-Type", contentType))
-              .build();
-      storage.create(blobInfo, Files.toByteArray(file));
+      s3Client.putObject(request);
     } else {
       LOG.info("dryrun - would have put file {} into s3 with key {}", file, key);
     }
     updatedKeys.add("/" + key);
   }
 
-  public static Builder builder(String projectId, String bucket) {
-    return new Builder(projectId, bucket);
+  public static Builder builder(String s3Bucket, String s3AccessKey, String s3SecretKey) {
+    return new Builder(s3Bucket, s3AccessKey, s3SecretKey);
   }
 
   /**
    * Builder to create the S3Publisher.
    */
   public static class Builder {
-
-    private final String projectId;
-    private final String bucket;
+    private final String s3Bucket;
+    private final String s3AccessKey;
+    private final String s3SecretKey;
     private String prefix;
+    private String cfDistribution;
+    private String cfAccessKey;
+    private String cfSecretKey;
     private boolean forcePush;
     private boolean dryrun;
+    private int timeout;
+    private Set<String> whitelist;
 
-    public Builder(String projectId, String bucket) {
-      this.projectId = projectId;
-      this.bucket = bucket;
-      this.forcePush = false;
-      this.dryrun = false;
-      this.prefix = "";
+    public Builder(String s3Bucket, String s3AccessKey, String s3SecretKey) {
+      this.s3Bucket = s3Bucket;
+      this.s3AccessKey = s3AccessKey;
+      this.s3SecretKey = s3SecretKey;
+      forcePush = false;
+      dryrun = false;
+      timeout = 30;
+      prefix = "";
+      whitelist = new HashSet<>();
+    }
+
+    public Builder setCloudfrontDistribution(String distribution) {
+      this.cfDistribution = distribution;
+      return this;
+    }
+
+    public Builder setCloudfrontAccessKey(String accessKey) {
+      this.cfAccessKey = accessKey;
+      return this;
+    }
+
+    public Builder setCloudfrontSecretKey(String secretKey) {
+      this.cfSecretKey = secretKey;
+      return this;
     }
 
     public Builder setPrefix(String prefix) {
@@ -232,8 +270,33 @@ public class GCSPublisher implements Publisher {
       return this;
     }
 
-    public GCSPublisher build() {
-      return new GCSPublisher(projectId, bucket, prefix, dryrun, forcePush);
+    public Builder setTimeout(int timeout) {
+      this.timeout = timeout;
+      return this;
+    }
+
+    public Builder setWhitelist(Set<String> whitelist) {
+      this.whitelist = whitelist;
+      return this;
+    }
+
+    public S3Publisher build() {
+      ClientConfiguration clientConf = new ClientConfiguration()
+        .withProtocol(Protocol.HTTPS)
+        .withSocketTimeout(timeout * 1000);
+
+      AmazonS3Client s3Client = new AmazonS3Client(new BasicAWSCredentials(s3AccessKey, s3SecretKey), clientConf);
+
+      AmazonCloudFrontClient cfClient = null;
+      if (cfDistribution != null) {
+        if (cfAccessKey == null || cfSecretKey == null) {
+          throw new IllegalArgumentException(
+            "When specifying a cloudfront distribution, must also specify a cloudfront access key and secret key.");
+        }
+        cfClient = new AmazonCloudFrontClient(new BasicAWSCredentials(cfAccessKey, cfSecretKey), clientConf);
+      }
+
+      return new S3Publisher(s3Client, cfClient, s3Bucket, prefix, cfDistribution, forcePush, dryrun, whitelist);
     }
   }
 }
